+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod io_utils;
-pub mod mapping;
-pub mod variables;
+pub mod processing;
+pub mod structures;
