+++ src/structures.rs
@@ -1,6 +1,35 @@
 use std::{iter::zip, ops::Index};
 
-use crate::mapping::TimerSV;
+pub const N_TALLIED_DATA: usize = 17;
+
+pub enum TalliedData {
+    Cycle = 0,
+    Start = 1,
+    Source = 2,
+    Rr = 3,
+    Split = 4,
+    Absorb = 5,
+    Scatter = 6,
+    Fission = 7,
+    Produce = 8,
+    Collision = 9,
+    Escape = 10,
+    Census = 11,
+    NumSeg = 12,
+    ScalarFlux = 13,
+    PopulationControl = 14,
+    CycleTracking = 15,
+    CycleSync = 16,
+}
+
+pub enum TimerSV {
+    Main = 0,
+    PopulationControl = 1,
+    CycleTracking = 2,
+    CycleTrackingKernel = 3,
+    CycleTrackingComm = 4,
+    CycleSync = 5,
+}
 
 #[derive(Debug)]
 pub struct FiniteDiscreteRV {
